requirements.txt
flask==3.1.1
anthropic>=0.45.0
openai>=1.0.0
rank-bm25==0.2.2
numpy>=1.26.0
requests>=2.31.0
faiss-cpu
fastembed
google-genai>=1.0.0
